requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.12.0
annotated-types==0.5.0
anyio==4.0.0
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.103.0
fastapi-mail==1.4.1
filelock==3.12.3
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
identify==2.5.27
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.15.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
langchain==0.0.279
langsmith==0.0.32
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nodeenv==1.8.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
orjson==3.9.5
packaging==23.1
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.16.1
pylint==2.17.5
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.4
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
